fungsi delete data komoditas

+++ app/Http/Controllers/KomoditasController.php
@@ -38,10 +38,13 @@ class KomoditasController extends Controller
     return redirect()->Route('komoditasData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
   }
 
-  public function Delete($Id){
-    $Id = HCrypt::Decrypt($Id);
-    $Komoditas = Komoditas::findOrFail($Id);
-    $Komoditas->delete();
-    return redirect()->Route('komoditasData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Delete Data Berhasil']);
+  public function Hapus($Id=null,$Verify=null){
+    if ($Verify) {
+      $Id = HCrypt::Decrypt($Id);
+      $Komoditas = Komoditas::findOrFail($Id);
+      $Komoditas->delete();
+      return redirect()->Route('komoditasData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Delete Data Berhasil']);
+    }
+    return abort(404);
   }
 }
